Update RankedGradeBook.cs
added a override on GetLetterGrade(). Now user can add grades via Ranked-grading system.

Ranked-grading grades students not based on their individual performance, but rather their performance compared to their peers. This means the 20% of the students with the highest grade of a class get an A, the next highest 20% get a B, etc.