Update StartingUserInterface.cs
In the StartingUserInterface class I refactered the CreateCommand(). The parts array now requires a 3rd string that will be the type. Next with a if sttatement the code creates a gradebook with the class that generates a new book of the corsepoding type.

I also edited the return text of HelpCommand() to let the user know they now need to  add a type