Added Weight Grading Support
Added a bool IsWeighted property to BaseGradeBook class. then refactored the constructors of the BaseGradeBook, StandardGradeBook, and RankedGradeBook classes require the IsWeighted property. Updated BaseGradeBook.getGPA() to add 1 to gpa when IsWeighted is true & the student type is Honors or dual enrolled. Finally I updated the StartingUserInterface.HelpCommand() changing the text to let user know to add wether if it’s a weighted grade